+++ playtime/src/models/json/track-json-store.js
@@ -47,6 +47,7 @@ export const trackJsonStore = {
 
   async deleteAllTracks() {
     await getDb().update((data) => {
+      /* eslint-disable-next-line no-param-reassign -- lowdb update() expects mutation */
       data.tracks = [];
     });
   },